backend/requirements.txt
requests
dotenv
flask
flask_cors
routes
flask_sqlalchemy
psycopg2
#psycopg2-binary
pandas
openai
schedule